prevent csrf attack against django integration by saving state
Previously, the integration was vulnerable against the kind of CSRF
attack described in RFC 6749 section 10.12. (See [1]). In this, a
user-agent could be forcibly logged into a new session by redirecting
it to a login-callback url that could be supplied by an attacker.

This has now been fixed by storing an authentication state in the
django session associated with a user-agent.

[1]: https://www.rfc-editor.org/rfc/rfc6749#section-10.12